Pass the address into fetchShippingMethodsEstimate
